A quick implementation of heapsort
This version of heapsort appears to run 1/3 as fast as STL quicksort.
It's also fairly ugly, but I managed to learn a little bit more about
passing function objects as template arguments (e.g. for the less
function object). I'd be interested to see what the standard deviation
of runtimes looks like for quicksort for small (10-1000) lists looks
like.